refactor: split monolithic identity.rs into module directory
Split src/identity.rs (1,351 lines) into src/identity/ with sub-modules:
- types.rs (150 lines): OperatingMode, MemoryPolicy
- contract.rs (342 lines): ManifestContract + validation helpers
- manifest.rs (257 lines): IdentityManifest + compose_with_precedence
- compose.rs (317 lines): IdentityManifest::compose impl
- loader.rs (363 lines): IdentityLoadRequest, IdentityLoader, BuiltinIdentityLoader
- mod.rs (11 lines): declarations + re-exports

All pub items re-exported from mod.rs so external callers are unaffected.
Updated tests/contracts.rs include_str! path to point at contract.rs.
